+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.2] - 2021-02-22
+
+* Added validateNucleotideSequence function.
+* Added countNucleotideFrequencyfunction.
+* Added generateRandomNucleotideSequence function.
+* Added transcribeDNAToRNA function.
+* Added generateComplementaryStrand function.
+* Added calculateGCContent function.
+
 ## [0.0.1] - 2021-02-21
 
 * Package created.